+++ CityOfBinds/src/game/binds/bind.py
@@ -1,4 +1,4 @@
-from ..utils.triggers import _Trigger, _TriggerMixin
+from ..utils import _Trigger, _TriggerMixin
 from ..utils.slash_commands import _CommandsMixin, _CommandGroup
