+++ CityOfBinds/src/game/utils/__init__.py
@@ -1,9 +1,15 @@
 from .comments.comment import _Comment
 from .comments.comment_banner import _CommentBanner
 from .powers.power import _Power
+from .triggers.trigger import _Trigger
+from .triggers.wasd_trigger import _WASDTrigger
+from .triggers.trigger_mixin import _TriggerMixin
 
 __all__ = [
     "_Comment",
     "_CommentBanner",
     "_Power",
+    "_Trigger",
+    "_WASDTrigger",
+    "_TriggerMixin",
 ]
