--- CityOfBinds/src/game/utils/triggers/__init__.py
@@ -1,5 +1,0 @@
-from .trigger import _Trigger
-from .wasd_trigger import _WASDTrigger
-from .trigger_mixin import _TriggerMixin
-
-__all__ = ["_Trigger", "_WASDTrigger", "_TriggerMixin"]
